update utils and add function execute cgi
