+++ app/cgi-bin/hello_world.py
@@ -1,8 +1,11 @@
 #!/usr/bin/python3
 
 def handle_call():
-	with open('../asset/index.html', 'r') as f:
-		print(f.read())
+	# print env
+	import os
+	print(os.environ)
+	
 
 if __name__ == '__main__':
 	handle_call()
+	exit(0)
