+++ app/includes/Webserv.hpp
@@ -12,16 +12,16 @@
 #include <vector>
 
 #include <sys/socket.h>
+#include <sys/types.h>
+#include <sys/wait.h>
 #include <netinet/in.h>
 #include <unistd.h>
+#include <netdb.h>
 
 #include "ConfigParser.hpp"
 #include "Constant.hpp"
 #include "Server.hpp"
 
-
-#include <sys/types.h>
-#include <netdb.h>
 
 #ifndef EXIT_FAILURE
 #define EXIT_FAILURE 1
@@ -34,4 +34,5 @@ int 						findCharInString(std::string str, const char *c);
 std::string 				ft_strtrim(std::string str);
 std::vector<std::string> 	ft_split(std::string str, const char *delimiter);
 
+std::string execute_cgi(int argc, char *argv[], char * const * envp);
 #endif
